포켓몬스터 수정 구현

+++ springhome/src/main/java/com/kh1/springhome/controller/PocketmonController.java
@@ -122,11 +122,8 @@ public class PocketmonController {
 //				.contentType(MediaType.APPLICATION_OCTET_STREAM)
 
 //				.header("Content-Disposition","attachment; filename="+attachDto.getAttachName())
-				.header(HttpHeaders.CONTENT_DISPOSITION,
-						ContentDisposition.attachment()
-								.filename(attachDto.getAttachName(), 
-										StandardCharsets.UTF_8).build().toString()
-								)
+				.header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
+						.filename(attachDto.getAttachName(), StandardCharsets.UTF_8).build().toString())
 				.body(resource);
 	}
 
@@ -152,12 +149,67 @@ public class PocketmonController {
 	public String delete(@RequestParam int no) {
 		AttachDto attachDto = pocketmonDao.findImage(no);
 		pocketmonDao.delete(no);// 포켓몬+이미지연결정보 삭제
-		attachDao.delete(attachDto.getAttachNo());// 파일정보 삭제
 
-		String home = System.getProperty("user.home");
-		File dir = new File(home, "upload");
-		File target = new File(dir, String.valueOf(attachDto.getAttachNo()));
-		target.delete();// 실제파일 삭제
+		//이미지 있는 경우 이미지 삭제 처리 추가
+		if (attachDto != null) {
+
+			String home = System.getProperty("user.home");
+			File dir = new File(home, "upload");
+			File target = new File(dir, String.valueOf(attachDto.getAttachNo()));
+			target.delete();// 실제파일 삭제
+			
+			attachDao.delete(attachDto.getAttachNo());// 파일정보 삭제
+		}
 		return "redirect:list";
+	}
+
+	// 수정
+	@GetMapping("/edit")
+	public String edit(@RequestParam int no, Model model) {
+		PocketmonDto pocketmonDto = pocketmonDao.selectOne(no);
+		model.addAttribute("pocketmonDto", pocketmonDto);
+		return "/WEB-INF/views/pocketmon/edit.jsp";
+	}
+
+	// 수정처리를 할 때 첨부파일 유무에 따라 다른 처리를 해야한다.
+	// - 첨부파일이 없으면 기존 첨부파일을 유지(아무것도 안하면됨)
+	// - 첨부파일이 있으면 기존 첨부파일을 변경(변경 대신 삭제+등록)
+	@PostMapping("/edit")
+	public String edit(@ModelAttribute PocketmonDto pocketmonDto, @RequestParam MultipartFile attach)
+			throws IllegalStateException, IOException {
+		// 포켓몬스터 정보 변경
+		pocketmonDao.update(pocketmonDto);
+
+		if (!attach.isEmpty()) {// 파일이 있으면
+			// 파일 삭제 - 기존 파일이 있을경우에만 처리
+			AttachDto attachDto = pocketmonDao.findImage(pocketmonDto.getNo());
+			String home = System.getProperty("user.home");
+				File dir = new File(home, "upload");
+				
+			if (attachDto != null) {
+				attachDao.delete(attachDto.getAttachNo());
+				
+				File target = new File(dir, String.valueOf(attachDto.getAttachNo()));
+				target.delete();
+			}
+			// 파일 추가 및 연결
+			int attachNo = attachDao.sequence();
+
+			File insertTarget = new File(dir, String.valueOf(attachNo));
+			attach.transferTo(insertTarget);
+
+			AttachDto insertDto = new AttachDto();
+
+			insertDto.setAttachNo(attachNo);
+			insertDto.setAttachName(attach.getOriginalFilename());
+			insertDto.setAttachSize(attach.getSize());
+			insertDto.setAttachType(attach.getContentType());
+			attachDao.insert(insertDto);
+			
+			pocketmonDao.connect(pocketmonDto.getNo(), attachNo);
+		}
+
+		return "redirect:/pocketmon/detail?no=" + pocketmonDto.getNo();
+
 	}
 }
